Refactor and Polishing — Stage 1

+++ src/main/java/com/aryan/studentmanagementapi/service/BranchService.java
@@ -7,11 +7,13 @@ import org.springframework.stereotype.Service;
 import com.aryan.studentmanagementapi.dto.BranchRequestDTO;
 import com.aryan.studentmanagementapi.dto.BranchResponseDTO;
 import com.aryan.studentmanagementapi.dto.BranchStudentCountDTO;
+import com.aryan.studentmanagementapi.exception.BadRequestException;
 import com.aryan.studentmanagementapi.exception.BranchAlreadyExistsException;
 import com.aryan.studentmanagementapi.exception.BranchNotFoundException;
 import com.aryan.studentmanagementapi.mapper.BranchMapper;
 import com.aryan.studentmanagementapi.model.Branch;
 import com.aryan.studentmanagementapi.repository.BranchRepository;
+import com.aryan.studentmanagementapi.repository.StudentRepository;
 
 import jakarta.transaction.Transactional;
 
@@ -19,10 +21,12 @@ import jakarta.transaction.Transactional;
 public class BranchService {
     
     private final BranchRepository branchRepository;
+    private final StudentRepository studentRepository;
     private final BranchMapper mapper;
 
-    public BranchService(BranchRepository branchRepository, BranchMapper mapper){
+    public BranchService(BranchRepository branchRepository, StudentRepository studentRepository, BranchMapper mapper){
         this.branchRepository = branchRepository;
+        this.studentRepository = studentRepository;
         this.mapper = mapper;
     }
 
@@ -47,6 +51,7 @@ public class BranchService {
             );
         
         Branch branch = mapper.toBranch(dto);
+        branchRepository.save(branch);
 
         return mapper.toBranchResponseDTO(branch);
     }
@@ -62,9 +67,11 @@ public class BranchService {
         branchRepository
             .findByName(dto.getName())
             .ifPresent(existingBranch -> {
-                    throw new BranchAlreadyExistsException(
-                        "Branch with name " + dto.getName() + " already exists." 
-                    );
+                    if(!existingBranch.getId().equals(branch.getId())){
+                        throw new BranchAlreadyExistsException(
+                            "Branch with name " + dto.getName() + " already exists." 
+                        );
+                    }
                 }
             );
         
@@ -79,6 +86,12 @@ public class BranchService {
         Branch branch = branchRepository
             .findByName(name)
             .orElseThrow(() -> new BranchNotFoundException(name));
+
+        if(studentRepository.existsByBranchId(branch.getId())){
+            throw new BadRequestException(
+                "Cannot delete brnach because students are assigned to it"
+            );
+        }
         
         branchRepository.delete(branch);
     }
